requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
bson==0.5.10
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.103.2
h11==0.14.0
idna==3.4
motor==3.3.1
mypy==1.6.0
mypy-extensions==1.0.0
passlib==1.7.4
pyasn1==0.5.0
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.5.0
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
slugify==0.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
